DBI methods

Implementations of pure virtual functions defined in the `DBI` package.

Usage

# S4 method for class 'connConnection,ANY'
dbWriteTable(conn, name, value, ...)

# S4 method for class 'connConnection'
dbSendQuery(conn, statement, ...)

Arguments

conn

A `DBIConnection-class` object, as returned by [dbConnect()].

name

The table name, passed on to dbQuoteIdentifier(). Options are:

a character string with the unquoted DBMS table name, e.g. "table_name",

a call to Id() with components to the fully qualified table name, e.g. Id(schema = "my_schema", table = "table_name")

a call to SQL() with the quoted and fully qualified table name given verbatim, e.g. SQL('"my_schema"."table_name"')

value

A `data.frame` (or coercible to `data.frame`).

...

Other parameters passed on to methods.

statement

a character string containing SQL.